pyproject.toml
[tool.poetry]
name = "orbitapi-backend"
version = "0.1.0"
description = ""
authors = ["abhishekY2401 <abhiyd2401@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.3"
pydantic = "^2.9.2"
uvicorn = "^0.32.0"
python-dotenv = "^1.0.1"
motor = "^3.6.0"
pydantic-settings = "^2.6.0"
google-generativeai = "^0.8.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
